requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
peft>=0.7.0
trl>=0.7.0
datasets>=2.14.0

# Audio processing
openai-whisper>=20231117
TTS>=0.22.0
pyaudio>=0.2.11
librosa>=0.10.0
soundfile>=0.12.0
pygame>=2.5.0

# LLM inference
llama-cpp-python>=0.2.0
sentence-transformers>=2.2.0

# Web interface
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Utilities
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
requests>=2.31.0
psutil>=5.9.0
python-dotenv>=1.0.0

# Development tools
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.7.0
pre-commit>=3.5.0

# Optional: Experiment tracking
wandb>=0.16.0
